Split Chat Into Its Own File/Lots and Lots of Bug Fixes
Chat is now sent to a file named 'chatlog.0', while all other log information is still sent to 'screenlog.0' as normal. Also included is a very large amount of bug fixes and optimizations, included the cause of the memory leak that was crashing the program the last time we tested.