(Hopefully) Final First Version of Discord Auth Handling
This is likely the last update for the Discord Auth. It now allows a
user to log in with Discord, checks if the user is a member of the
server, then checks if the user has the correct roles. It also allows
for two levels of access. The only thing left to do with this code is to
add the redirects and username/level assignment.