Server Settings Configuration
This is now completed. There is probably some some things to be done to
make html formatting prettier for a few things, mostly on returned
output. But it is functioning. Settings can be updated via web control
or direct url. Settings changes are logged in a file. javascript and php
validation works. Not all settings have to be updated at the same time
when direct URL is used.