Session end changes
Found why php hangs. It's per user session. Since we use sessions to
handle user authentication, we open the session on every page. While the
session is open, php will not allow other sessions to open (makes
sense). So now we should be closing the session when it's not needed.