add server version selection
Try one for adding server version selection. Added searching for version
folders, and a drop down to select the vesrion for each server.

Added server-settings-web.json. If not found, it will be created
automatically when viewing server configs. Will need to check and create
again when server start accepts the variable.