server path to manage.c
factorio versions should now be stored in
/usr/share/factorio/VERSION_NUMBER/

the version to start the server on can be configured from config for
each server

version selection data is stored in server-settings-web.json in each
servers data folder. the file or variable do not exist, it will be
generated when viewing the config page.

the first folder in /usr/share/factorio/ will be selected as the default
when the server-settings-web.json file is being generated.

when starting a server, the server-settings-web.json file will be called
to check if version is set. This version will be checked against the
available versions in the /usr/share/factorio folder. If the version is
valid (or if the folder exists) the folder will be passed into
manage.sh. manage.sh will send the path and the character count or
string length of the path to the manage.c

now we just need to update manage.c, and the web control will support
multiple versions of factorio.