version manager updates
displaying in a table and adding some buttons.

still needed:
backend installing work. I might have to create another bash file for
this, so it can be run in the background. But I may also be able to run
a php script in the background, which would be nicer to reduce file
counts.

javascript to send install/delete commands for server backend use.