Crontab install fix
Wasn’t installing because there was no new line at the end of the file.
Come on unix! Be smarter than this!

Also added a pause in the install to inform the user that their old
crontab will be saved to a file in case there was anything important to
them in there, since this will replace the entire crontab job list.