login and altlogin combine
login and altlogin are now the same. If specific post data is found, the
alternate login will be displayed. If the clientid in config.json is not
found or is a default setting, then the alternate login will also
display.

Either login can be easily accessed using a show/hide link.