bugfix memory display color
Math was bad on the memory calculation. This should work now. Given, this is going purely on a percentage. Better logic should be worked on to know if memory is actually low (ie, if 64gb out of 128gb RAM is utilizaed on a server, that 50% utilization is going to change the color from green when there is plenty of headroom left.)