Pull New Bans from Global Banlist
Every X minutes (configurable in ban_config.json), the program will request to download the most recent bans from the global banlist. This will keep the banlist in sync with the other organizations running the program.
The function is currently coded to not run, as the endpoint for retrieving the most recent bans is not yet fully functional.